pyproject.toml
[tool.poetry]
name = "alloff-backoffice-server"
version = "0.1.0"
description = ""
authors = ["Steve Cheong <j.stephen.cheong@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
djangorestframework = "^3.12.4"
Markdown = "^3.3.4"
django-filter = "^21.1"
mongoengine = "^0.23.1"
iamport-rest-client = "^0.8.2"
django-cors-headers = "^3.8.0"
drf-yasg = "^1.20.0"
django-rest-framework-mongoengine = {git = "git@github.com:lessbutter/django-rest-framework-mongoengine.git"}


[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"
black = "^21.9b0"
PyYAML = "^5.4.1"
uritemplate = "^3.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
